requirements.txt
pandas~=1.2.1
requests~=2.22.0
Pillow~=8.1.0
pytest~=6.2.2